autcv/backend_cv/requirements.txt
fastapi[standard]
groq
python-dotenv
fpdf
requests
